+++ tests/test_model.py
@@ -160,29 +160,30 @@ def test_CouplingModel_explicit():
     W1_new = H_mpo.get_W(1)
     Id, JW, N = fermion_site.Id, fermion_site.JW, fermion_site.N
     Cd, C = fermion_site.Cd, fermion_site.C
-    CdJW = Cd.matvec(JW)
-    JWC = JW.matvec(C)
+    CdJW = Cd.matvec(JW) # = Cd
+    CJW = C.matvec(JW) # = -C
     # yapf: disable
-    W0_ex = [[Id,   None, None, CdJW, JWC,  N,    None, None, None, N*0.125],
+    W0_ex = [[Id,   None, None, CJW, CdJW,  N,    None, None, None, N*0.125],
              [None, None, Id,   None, None, None, None, None, None, None],
              [None, None, None, None, None, None, None, None, None, N*4.0],
+             [None, None, None, None, None, None, None, None, None, Cd*-1.5],
              [None, None, None, None, None, None, None, None, None, C*1.5],
-             [None, None, None, None, None, None, None, None, None, Cd*1.5],
              [None, Id,   None, None, None, None, None, None, None, None],
              [None, None, None, None, None, None, JW,   None, None, None],
              [None, None, None, None, None, None, None, JW,   None, None],
              [None, None, None, None, None, None, None, None, Id,   None],
              [None, None, None, None, None, None, None, None, None, Id]]
-    W1_ex = [[Id,  None, None, None, None, None,  CdJW, JWC,  N,    N*0.125],
+    W1_ex = [[Id,  None, None, None, None, None,  CJW,  CdJW, N,    N*0.125],
              [None, Id,   None, None, None, None, None, None, None, None],
              [None, None, None, None, None, None, None, None, None, N*4.0],
-             [None, None, None, JW,   None, None, None, None, None, C*0.5],
-             [None, None, None, None, JW,   None, None, None, None, Cd*0.5],
+             [None, None, None, JW,   None, None, None, None, None, Cd*-0.5],
+             [None, None, None, None, JW,   None, None, None, None, C*0.5],
              [None, None, None, None, None, Id,   None, None, None, None],
+             [None, None, None, None, None, None, None, None, None, Cd*-1.5],
              [None, None, None, None, None, None, None, None, None, C*1.5],
-             [None, None, None, None, None, None, None, None, None, Cd*1.5],
              [None, None, Id,   None, None, None, None, None, None, None],
              [None, None, None, None, None, None, None, None, None, Id]]
+
     # yapf: enable
     W0_ex = npc.grid_outer(W0_ex, W0_new.legs[:2])
     W1_ex = npc.grid_outer(W1_ex, W1_new.legs[:2])
@@ -208,31 +209,33 @@ def test_MultiCouplingModel_explicit():
     W2_new = H_mpo.get_W(2)
     Id, JW, N = fermion_site.Id, fermion_site.JW, fermion_site.N
     Cd, C = fermion_site.Cd, fermion_site.C
-    CdJW = Cd.matvec(JW)
-    JWC = JW.matvec(C)
+    CdJW = Cd.matvec(JW)  # = Cd
+    CJW = C.matvec(JW) # = -C
     NJW = N.matvec(JW)
+    import pprint
+    pprint.pprint(M.H_MPO_graph._build_grids())
     # yapf: disable
-    W0_ex = [[Id,   None, None, CdJW, None, JWC,  N,    None, None, None, N*0.125],
+    W0_ex = [[Id,   None, None, CJW,  CdJW, None, N,    None, None, None, N*0.125],
              [None, None, Id,   None, None, None, None, None, None, None, None],
              [None, None, None, None, None, None, None, None, None, None, N*4.0],
+             [None, None, None, None, None, None, None, None, None, None, Cd*-1.5],
              [None, None, None, None, None, None, None, None, None, None, C*1.5],
-             [None, None, None, None, JW,   None, None, None, None, None, None],
-             [None, None, None, None, None, None, None, None, None, None, Cd*1.5],
+             [None, None, None, None, None, JW,   None, None, None, None, None],
              [None, Id,   None, None, None, None, None, None, None, None, None],
              [None, None, None, None, None, None, None, None, None, None, C*1.125],
              [None, None, None, None, None, None, None, JW,   None, None, None],
              [None, None, None, None, None, None, None, None, JW,   None, None],
              [None, None, None, None, None, None, None, None, None, Id,   None],
              [None, None, None, None, None, None, None, None, None, None, Id]]
-    W1_ex = [[Id,   None, None, None, None, None, None, None, CdJW, JWC,  N,    N*0.125],
+    W1_ex = [[Id,   None, None, None, None, None, None, None, CJW,  CdJW, N,    N*0.125],
              [None, Id,   None, None, None, None, None, None, None, None, None, None],
              [None, None, None, None, None, None, None, None, None, None, None, N*4.0],
-             [None, None, None, JW,   NJW,  None, None, None, None, None, None, C*0.5],
+             [None, None, None, JW,   None, None, None, None, None, None, None, Cd*-0.5],
+             [None, None, None, None, JW,   NJW,  None, None, None, None, None, C*0.5],
              [None, None, None, None, None, None, None, None, None, None, None, C*1.125],
-             [None, None, None, None, None, JW,   None, None, None, None, None, Cd*0.5],
              [None, None, None, None, None, None, Id,   CdJW, None, None, None, None],
+             [None, None, None, None, None, None, None, None, None, None, None, Cd*-1.5],
              [None, None, None, None, None, None, None, None, None, None, None, C*1.5],
-             [None, None, None, None, None, None, None, None, None, None, None, Cd*1.5],
              [None, None, Id,   None, None, None, None, None, None, None, None, None],
              [None, None, None, None, None, None, None, None, None, None, None, Id]]
     # yapf: enable